# packages/contracts/openapi/control-plane.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpsAtlas control plane",
    "description" : "The control plane for the OpsAtlas service catalog.\n\nConventions that hold for every endpoint: collections are cursor-paginated and never offset-paginated; failures are RFC 9457 problem documents carrying a correlationId and, for validation failures, a violations array locating each problem by JSON Pointer; a correlation id is accepted on X-Correlation-Id or generated, and echoed on every response.\n\nEvery endpoint requires a caller. A person presents a bearer token from the configured identity provider; a machine of this system's own - the observer - presents the key it was issued on X-OpsAtlas-Key. A token that verifies but belongs to nobody provisioned here is answered 403, not 401: the credential is genuine, it is simply not known to this deployment.\n\nHealth here is probe availability, measured by an observer against a declared health endpoint from one vantage point. It is not an SLO measurement, and no field in this API should be read as one.\n",
    "version" : "v1"
  },
  "security" : [ {
    "bearerToken" : [ ]
  }, {
    "serviceCredential" : [ ]
  } ],
  "paths" : {
    "/api/v1/services/{slug}" : {
      "get" : {
        "tags" : [ "service-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceDetail"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "service-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "sourceRef",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/yaml" : {
              "schema" : {
                "type" : "string"
              }
            },
            "text/yaml" : {
              "schema" : {
                "type" : "string"
              }
            },
            "application/x-yaml" : {
              "schema" : {
                "type" : "string"
              }
            },
            "text/plain" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "service-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/services" : {
      "get" : {
        "tags" : [ "service-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "cursor",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 25,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseServiceSummary"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "service-controller" ],
        "operationId" : "register",
        "parameters" : [ {
          "name" : "sourcePath",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "service.yaml",
            "pattern" : "^[A-Za-z0-9._-]+(/[A-Za-z0-9._-]+)*$"
          }
        }, {
          "name" : "sourceRef",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/yaml" : {
              "schema" : {
                "type" : "string"
              }
            },
            "text/yaml" : {
              "schema" : {
                "type" : "string"
              }
            },
            "application/x-yaml" : {
              "schema" : {
                "type" : "string"
              }
            },
            "text/plain" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "EnvironmentView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "readinessPath" : {
            "type" : "string"
          },
          "livenessPath" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "livenessPath", "name", "readinessPath", "url" ]
      },
      "JsonNode" : {
        "type" : "object"
      },
      "ServiceDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "slug" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "repository" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "integer",
            "format" : "int32"
          },
          "runtime" : {
            "type" : "string"
          },
          "lifecycle" : {
            "type" : "string"
          },
          "owner" : {
            "type" : "string"
          },
          "schemaVersion" : {
            "type" : "string"
          },
          "manifestDigest" : {
            "type" : "string"
          },
          "sourcePath" : {
            "type" : "string"
          },
          "sourceRef" : {
            "type" : "string"
          },
          "environments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EnvironmentView"
            }
          },
          "manifest" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "registeredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "displayName", "environments", "id", "lifecycle", "manifest", "manifestDigest", "owner", "registeredAt", "repository", "runtime", "schemaVersion", "slug", "sourcePath", "sourceRef", "tier", "updatedAt", "version" ]
      },
...
      "PageResponseServiceSummary" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ServiceSummary"
            }
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "required" : [ "items", "nextCursor" ]
      },
      "ServiceSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "slug" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "repository" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "integer",
            "format" : "int32"
          },
          "runtime" : {
            "type" : "string"
          },
          "lifecycle" : {
            "type" : "string"
          },
          "owned" : {
            "type" : "boolean"
          },
          "registeredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "displayName", "id", "lifecycle", "owned", "registeredAt", "repository", "runtime", "slug", "tier", "version" ]
      },
...
    },
    "securitySchemes" : {
      "bearerToken" : {
        "type" : "http",
        "description" : "A token from the configured identity provider. Verified against its published key set; this system holds no signing key and cannot mint one.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "serviceCredential" : {
        "type" : "apiKey",
        "description" : "A key issued by this system to one of its own machines. The audit log records the caller as service:<name> rather than as a person.",
        "name" : "X-OpsAtlas-Key",
        "in" : "header"
      }
    }
  }
}